pyproject.toml
[tool.poetry]
name = "fastapicore"
version = "0.1.0"
description = ""
authors = ["sakthiRatnam <sakthiratnam050@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9"
fastapi = "^0.68.1"
FastAPI-SQLAlchemy = "0.2.1"
SQLAlchemy = "1.4.18"
PyJWT = "^2.1.0"
websockets = "^10.0"
tortoise-orm = "^0.17.8"
alembic = "1.6.5"
aerich = "^0.5.8"
pymongo = {extras = ["srv"], version = "^3.12.0"}
mongoengine = "0.23.1"
motor = "2.5.0"
toml = "0.10.2"
aiofiles = "0.7.0"
uvicorn = "0.13.4"
uvloop = "0.15.2"
asyncpg = "0.23.0"
aiosqlite = "0.16.1"
reportlab = "3.5.68"
emails = "0.6"
Jinja2 = "^3.0.1"
pydantic = "^1.8.2"
graphene = "2.1.8"
python-multipart = "^0.0.5"
tortoise-cli = "^0.1.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
py-bcrypt = "^0.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
